Solve a bug in shader code: input and output signature doesn't match

+++ D3D11Playground/shaders/matrix_mul_cs.hlsl
@@ -1,6 +1,6 @@
 cbuffer Parameters : register(b0)
 {
-	int kSize;
+	uint kSize;
 	int3 unused;
 }
 
@@ -11,11 +11,11 @@ RWStructuredBuffer<float> result : register(u0);
 [numthreads(256, 1, 1)]
 void main(int3 dispatch_thread_id: SV_DispatchThreadID)
 {
-	int i = dispatch_thread_id.x / kSize;
-	int j = dispatch_thread_id.x % kSize;
+	uint i = uint(dispatch_thread_id.x) / kSize;
+	uint j = uint(dispatch_thread_id.x) % kSize;
 
 	float temp = 0;
-	for (int k = 0; k < kSize; ++k)
+	for (uint k = 0; k < kSize; ++k)
 		temp += input[i * kSize + k] * input[k * kSize + j];
 	result[dispatch_thread_id.x] = temp;
 }
